pyproject.toml
[project]
name = "oscioler"
requires-python = ">=3.8"
authors = [
    {name = "Leonardo Mascelli"},
    {email = "lmascelli@gmail.com"},
]
mantainers = [
    {name = "Leonardo Mascelli", email = "lmascelli@gmail.com"},
]
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "numpy",
    "matplotlib",
    "scipy",
	"pyvisa",
	"pyvisa_py",
	"psutil",
	"zeroconf",
	"pyserial",
]